+++ src/components/menu/MenuListBlock.js
@@ -113,6 +113,11 @@ const MenuListBlock = (props) => {
         props.onClickDispatch(menuCode);
     }
     const onClickDelete = (e, menuCode) => {
+        let confirmRtn = window.confirm("删除是不可恢复的，确认要删除吗？");
+        if (!confirmRtn) {
+            return;
+        }
+
         let url = genGetUrlBySegs('/menuset/menu/{segment}/{segment}/delete', [getTenantCode(), menuCode]);
         axios.delete(url, {
             // withCredentials: true // 这会让axios在请求中携带cookies
